Move all objecttypes to fields.py and remove old objects
